requirements.txt
backports.zoneinfo==0.2.1
certifi==2021.10.8
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pip==21.2.4
psycopg==3.0.12
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
scipy==1.8.0
setuptools==61.2.0
six==1.16.0
SQLAlchemy==1.4.36
wheel==0.37.1
